#10 refactor : Comment
메소드 변경점 test코드에 추가
card쪽 필요하기에 아직 테스트 불가
