+++ src/test/java/com/sparta/jamrello/domain/comment/service/CommentService/CommentServiceTest.java
@@ -7,6 +7,7 @@ import com.sparta.jamrello.domain.card.repository.entity.Card;
 import com.sparta.jamrello.domain.comment.dto.CommentRequestDto;
 import com.sparta.jamrello.domain.comment.repository.CommentRepository;
 import com.sparta.jamrello.domain.comment.repository.entity.Comment;
+import com.sparta.jamrello.domain.comment.service.CommentService;
 import com.sparta.jamrello.domain.member.repository.entity.Member;
 import com.sparta.jamrello.domain.member.repository.entity.MemberRepository;
 import com.sparta.jamrello.global.exception.BisException;
@@ -15,7 +16,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import static org.mockito.BDDMockito.*;
 import java.util.Optional;
